enforce data types for all data structures in Mesh

+++ mesh.py
@@ -47,7 +47,8 @@ class Mesh:
         return self._admittances
 
     def _calc_corners(self):
-        corners = np.empty((*self._elements.shape, self._nodes.shape[1]))
+        corners = np.empty((*self._elements.shape, self._nodes.shape[1]),
+                           dtype=float)
         for i, corner_indexes in enumerate(self._elements):
             corners[i] = np.vstack([self._nodes[j] for j in corner_indexes])
         return corners
@@ -56,10 +57,11 @@ class Mesh:
         return self._corners.mean(axis=1)
 
     def _calc_normals(self):
-        normals = np.empty((self._elements.shape[0], self._nodes.shape[1]))
+        normals = np.empty((self._elements.shape[0], self._nodes.shape[1]),
+                           dtype=float)
         for i, corners in enumerate(self._corners):
             element_vector = corners[1] - corners[0]
             element_length = np.sqrt(element_vector.dot(element_vector))
             normals[i] = element_vector.dot(
-                np.array([[0, -1], [1, 0]]))/element_length
+                np.array([[0., -1.], [1., 0.]]))/element_length
         return normals
